Disables GZip compression for jabsorb
This sets the GZip threshold to -1, effectively disabling gzip
compression for every json response.

The jabsorb GZip threshold is the minimum number of bytes gained from
compression that makes acceptable to send a compressed json to the
client. When the result of the compression falls below the threshold,
the compressed message is discarded and the original one is sent
instead.
Jemma's compressed json messages are usually so small that the
compressed message is actually bigger than the original one, so ATM
compressing them seems useless.